io_util

Module batch

Helper traits and types for asynchronous I/O.

The IoCompletion trait abstracts one asynchronous I/O operation for which a task can call IoCompletion::wait. A list of IoCompletion objects can be held by an IoBatch, which provides an IoBatch::wait_all method so that a task can wait until the whole batch of asynchronous I/O operations completes.

Examples

The common pattern is to pass an IoBatch as an out-parameter when submitting I/O. The submission path pushes a completion record only when the operation will complete asynchronously. The caller then waits for all records that were added to the batch.

use std::sync::Arc;

use io_util::{
    IoError,
    batch::{IoBatch, IoCompletion},
};

struct ReadCompletion {
    result: Result<(), IoError>,
}

impl IoCompletion for ReadCompletion {
    fn wait(&self) -> Result<(), IoError> {
        self.result
    }
}

struct MockDevice;

impl MockDevice {
    fn submit_read(&self, io_batch: &mut IoBatch) -> Result<(), IoError> {
        io_batch.push(Arc::new(ReadCompletion { result: Ok(()) }));
        Ok(())
    }

    fn submit_fail_read(&self, io_batch: &mut IoBatch) -> Result<(), IoError> {
        io_batch.push(Arc::new(ReadCompletion {
            result: Err(IoError::Failed),
        }));
        Ok(())
    }
}

let device = MockDevice;
let mut io_batch = IoBatch::new();
device.submit_read(&mut io_batch)?;
assert_eq!(io_batch.len(), 1);
io_batch.wait_all()?;

let mut io_batch = IoBatch::new();
device.submit_fail_read(&mut io_batch)?;
assert_eq!(io_batch.len(), 1);
assert_eq!(io_batch.wait_all(), Err(IoError::Failed));

Structs

IoBatch

A batch of IoCompletion records.

Traits

IoCompletion

A handle to one asynchronous I/O operation.